progreso

+++ Practica4/IG1App/Scene0.cpp
@@ -11,8 +11,8 @@ void Scene0::init()
 
 	// Graphics objects (entities) of the scene
 	gObjects.push_back(new RGBAxes(400.0));
-	gObjects.push_back(new RegularPolygon(3, 100, glm::dvec4(0, 255, 255, 1)));
-	gObjects.push_back(new RegularPolygon(100, 100, glm::dvec4(255, 0, 255, 1)));
+	gObjects.push_back(new RegularPolygon(3, 100, glm::vec4(0, 255, 255, 1)));
+	gObjects.push_back(new RegularPolygon(100, 100, glm::vec4(255, 0, 255, 1)));
 	
 	RGBTriangle* rgbTriangle = new RGBTriangle(20);
