+++ Practica4/IG1App/Scene2.cpp
@@ -13,7 +13,7 @@ void Scene2::init()
 	gObjects.push_back(new RGBAxes(400.0));
 	gObjects.push_back(new RGBTriangle(20));
 	gObjects.push_back(new RGBRectangle(400, 200));
-	gObjects.push_back(new RegularPolygon(100, 200, glm::dvec4(1)));
+	gObjects.push_back(new RegularPolygon(100, 200, glm::vec4(1)));
 }
 
 void Scene2::update()
